redesigned the project architecture
