+++ app.js
@@ -1,13 +1,12 @@
-import { initMap, createTransportContainer, registerUser } from "./src/index.js";
-import { logInUser } from "./src/index.js";
+import { initMap, createTransportContainer, registerUser, logInUser } from "./src/index.js";
 
-// const userId = localStorage.getItem('userId');
+const userId = localStorage.getItem('userId');
 
-// if (!userId) {
-//     registerUser()  
-// } else {
-//     initMap();
-//     createTransportContainer();
-// }
+if (!userId) {
+    registerUser()  
+} else {
+    initMap();
+    createTransportContainer();
+}
 
-console.log(logInUser())
+logInUser()
