+++ src/forms/logIn/logIn.html
@@ -1,9 +1,19 @@
-import { ROOT_ELEMENT } from "../constants";
+<!DOCTYPE html>
+<html lang="en">
 
-export function drawLogInForm() {
-    ROOT_ELEMENT.insertAdjacentHTML('beforeend', `
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Log In</title>
+    <link rel="stylesheet" href="../style.css">
+    <link
+        href="https://fonts.googleapis.com/css2?family=Montserrat+Alternates:ital,wght@0,500;0,600;0,700;0,800;1,500;1,600;1,700;1,800&display=swap"
+        rel="stylesheet">
+</head>
+
+<body>
     <div class="container">
-        <form action="" method="GET" class="form" name="logIn">
+        <form action="#" method="GET" class="form">
             <h2 class="title">RentRide</h2>
             <p class="subtitle">Log In</p>
 
@@ -20,5 +30,7 @@ export function drawLogInForm() {
             <button type="submit">Sign In</button>
             <p class="login-text">No account? <a href="#" class="login-link">Create an account</a></p>
         </form>
-    </div>`)
-}
+    </div>
+</body>
+
+</html>
