+++ src/index.js
@@ -1,7 +1,7 @@
 export { firebase } from "./firebaseActions.js";
 export { initMap } from './map/initMap.js';
 export { createTransportContainer } from './transportContainer/transportContainer.js';
-export { registerUser } from './forms/forms.js';
-export { logInUser } from './forms/forms.js'
+export { registerUser } from './forms/register/registerUser.js';
+export { logInUser } from './forms/logIn/logInUser.js';
 export { encryptPassword } from './encryption.js';
 export { decryptPassword } from './encryption.js'
